refactored daemon check into validateNode
